// webview-ui/src/thor/model/SalesActivity.ts
export type SalesActivity = {
  /**
   * The associated opportunity.
   * @type {string}
   * @memberof SalesActivity
   */
  opportunityId: string;
  /**
   * Type of the sales activity.
   * @type {string}
   * @memberof SalesActivity
   */
  type: SalesActivityTypeEnum;
  /**
   * Date and time of the activity.
   * @type {Date}
   * @memberof SalesActivity
   */
  activityDate: Date;
  /**
   * The user who performed the activity.
   * @type {string}
   * @memberof SalesActivity
   */
  performedBy: string;
  /**
   * Details about the activity.
   * @type {string}
   * @memberof SalesActivity
   */
  details?: string;
  /**
   * Unique identifier for object in the system
   * @type {string}
   * @memberof SalesActivity
   */
  id?: string;
  /**
   * UUID of owner of the object in the system
   * @type {string}
   * @memberof SalesActivity
   */
  ownerId?: string;
  /**
   * Date of object creation
   * @type {Date}
   * @memberof SalesActivity
   */
  createdDate?: Date;
  /**
   * Data, including hash of the key(s) used to encrypt this record.
   * @type {string}
   * @memberof SalesActivity
   */
  keyHash?: string;
  /**
   * Last user to access object
   * @type {string}
   * @memberof SalesActivity
   */
  lastAccessedById?: string;
  /**
   * Timestamp of last access of object
   * @type {Date}
   * @memberof SalesActivity
   */
  lastAccessedDate?: Date;
  /**
   * Unique identifier for user who last modifed the object in the system
   * @type {string}
   * @memberof SalesActivity
   */
  lastModifiedById?: string;
  /**
   * Date of last object modification
   * @type {Date}
   * @memberof SalesActivity
   */
  lastModifiedDate?: Date;
};
// ...
export enum SalesActivityTypeEnum {
  CALL = "call",
  MEETING = "meeting",
  EMAIL = "email",
  FOLLOWUP = "follow-up",
}